requirements.txt
scapy
pydivert
netfilterqueue
python-nmap